distracted left & right side; finished left side
