+++ src/components/GameConfigCard.tsx
@@ -1,0 +1,15 @@
+import { CardBoard } from "../containers/Home";
+
+const GameConfigCard: React.FunctionComponent = () => {
+
+  return (
+    <CardBoard>
+      <h2>Game config</h2>
+      <wired-card>
+        
+      </wired-card>
+    </CardBoard>
+  )
+}
+
+export default GameConfigCard;
